# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Revenge API",
    "description": "الانتقام (Revenge) — Auth, Game State, and Leaderboard API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/v1/auth/login or /api/v1/auth/register"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "البريد الإلكتروني غير صحيح" }
        },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "username": { "type": "string", "example": "ashraf" },
          "email": { "type": "string", "format": "email", "example": "ashraf@example.com" },
          "createdAt": { "type": "string", "format": "date-time" },
          "last_login_at": { "type": "string", "format": "date-time" },
          "leaderboard_opt_out": { "type": "integer", "example": 0 }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "example": "ashraf"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ashraf@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Must contain at least one letter and one number",
            "example": "TestPass123!"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "ashraf@example.com" },
          "password": { "type": "string", "example": "TestPass123!" }
        }
      },
...
      "GameSave": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "slot": { "type": "integer", "example": 1 },
          "heroLevel": { "type": "integer", "example": 5 },
          "heroXp": { "type": "integer", "example": 1200 },
          "currentMap": { "type": "string", "example": "forest" },
          "totalKills": { "type": "integer", "example": 42 },
          "totalDistance": { "type": "number", "example": 1500.5 },
          "completedGame": { "type": "boolean", "example": false },
          "heroStats": { "type": "object", "additionalProperties": true },
          "inventory": { "type": "object", "additionalProperties": true },
          "progress": { "type": "object", "additionalProperties": true },
          "worldState": { "type": "object", "nullable": true, "additionalProperties": true },
          "meta": { "type": "object", "additionalProperties": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SaveGameRequest": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "heroStats": { "type": "object", "additionalProperties": true },
          "inventory": { "type": "object", "additionalProperties": true },
          "progress": {
            "type": "object",
            "properties": {
              "heroLevel": { "type": "integer", "example": 5 },
              "currentMap": { "type": "string", "example": "forest" },
              "completedGame": { "type": "boolean", "example": false }
            }
          },
          "worldState": { "type": "object", "additionalProperties": true },
          "meta": {
            "type": "object",
            "properties": {
              "totalKills": { "type": "integer", "example": 42 },
              "totalDistance": { "type": "number", "example": 1500.5 }
            }
          }
        }
      },
      "SyncRequest": {
        "type": "object",
        "required": ["saves"],
        "properties": {
          "saves": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["slot", "data"],
              "properties": {
                "slot": { "type": "integer", "example": 1 },
                "timestamp": { "type": "string", "format": "date-time" },
                "data": { "$ref": "#/components/schemas/SaveGameRequest" }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email or username already taken", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "423": { "description": "Account locked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/game/sync": {
      "post": {
        "tags": ["Game Saves"],
        "summary": "Bulk sync multiple save slots",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SyncRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sync result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saves": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slot": { "type": "integer" },
                          "data": { "$ref": "#/components/schemas/GameSave" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid payload", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}